requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.2
PyMuPDF==1.23.9
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0